common, kernel, dt: break the circular import via uenv.py
- Move get_image_bootloader() from deploy.py to common.py.
- Move get_current_uenv_txt_path(), set_uenv_txt_variable() and
  get_uenv_txt_variable() from dt.py to uenv.py.
- Replace separate get_dt_changes_dir()/get_kernel_changes_dir()
  functions in dt.py and kernel.py with get_changes_dir(subdir). dt.py,
  kernel.py and uenv.py call get_changes_dir(subdir) to get their
  respective changes directory.
